+++ plugins/prometheus/src/components/page/PageChart.tsx
@@ -86,8 +86,8 @@ const PageChart: React.FunctionComponent<IPageChartProps> = ({ queries, times, m
             }}
             // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
             tooltip={(tooltip) => {
-              const serie = selectedSeries.filter((serie) => serie.id === tooltip.point.serieId);
-              const isFirstHalf = tooltip.point.index % serie[0].data.length < serie[0].data.length / 2;
+              const isFirstHalf =
+                Math.floor(new Date(tooltip.point.data.x).getTime() / 1000) < (times.timeEnd + times.timeStart) / 2;
 
               return (
                 <TooltipWrapper anchor={isFirstHalf ? 'right' : 'left'} position={[0, 20]}>
@@ -111,7 +111,7 @@ const PageChart: React.FunctionComponent<IPageChartProps> = ({ queries, times, m
                 </TooltipWrapper>
               );
             }}
-            xScale={{ type: 'time' }}
+            xScale={{ max: new Date(times.timeEnd * 1000), min: new Date(times.timeStart * 1000), type: 'time' }}
             yScale={{ stacked: stacked, type: 'linear' }}
             yFormat=" >-.2f"
           />